registration

+++ app/Http/Controllers/TaskController.php
@@ -14,10 +14,20 @@ class TaskController extends Controller
      */
     public function index()
     {
-        $tasks = Task::all();
-        return view ('tasks.index', [
-            'tasks' => $tasks
-            ]);
+        $tasks = [];
+         if (\Auth::check()) { 
+             $user = \Auth::user();
+             $tasks = $user->tasks()->orderBy('created_at', 'asc')->paginate(10);
+         
+             
+         }
+            return view('tasks.index', [
+                'tasks' => $tasks,
+                ]);
+            
+         
+        
+    
     }
 
     /**
@@ -45,10 +55,11 @@ class TaskController extends Controller
             'status' => 'required|max:10',
             ]);
             
-        $task = new Task;
-        $task->content = $request->content;
-        $task->status = $request->status;
-        $task->save();
+        $request->user()->tasks()->create([
+            'content' => $request->content,
+            'status' => $request->status,
+        ]);
+
         
         return redirect('/');
     }
@@ -113,6 +124,7 @@ class TaskController extends Controller
     public function destroy($id)
     {
         $task = Task::findOrFail($id);
+        if (\Auth::id() === $task->user_id)
         $task->delete();
         
         return redirect('/');
